Add DispatcherTimer wrapper class PomodoroTimer and start code refactor around that
